Move music player functionality to voice player module
