+++ app/src/main/java/ir/mehdiyari/krypt/ui/voice/audios/AudiosListComponents.kt
@@ -29,9 +29,9 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.constraintlayout.compose.ConstraintLayout
 import ir.mehdiyari.krypt.R
-import ir.mehdiyari.krypt.ui.voice.audios.entity.AudioEntity
-import ir.mehdiyari.krypt.ui.voice.player.MusicPlayerEntity
 import ir.mehdiyari.krypt.core.designsystem.theme.KryptTheme
+import ir.mehdiyari.krypt.voice.player.entity.MusicPlayerEntity
+import ir.mehdiyari.krypt.voice.shared.entity.AudioEntity
 
 @Composable
 fun AudioList(
