+++ app/src/main/java/ir/mehdiyari/krypt/ui/voice/audios/entity/AudioEntityMapper.kt
@@ -3,8 +3,9 @@ package ir.mehdiyari.krypt.ui.voice.audios.entity
 import ir.mehdiyari.krypt.file.data.entity.FileEntity
 import ir.mehdiyari.krypt.ui.voice.audios.AudioTimeMapper
 import ir.mehdiyari.krypt.ui.voice.recorder.SecondToTimerMapper
-import ir.mehdiyari.krypt.ui.voice.recorder.meta.AudioMetaData
-import ir.mehdiyari.krypt.ui.voice.recorder.meta.AudioMetaDataJsonParser
+import ir.mehdiyari.krypt.voice.shared.entity.AudioEntity
+import ir.mehdiyari.krypt.voice.shared.entity.meta.AudioMetaData
+import ir.mehdiyari.krypt.voice.shared.entity.meta.AudioMetaDataJsonParser
 import javax.inject.Inject
 
 class AudioEntityMapper @Inject constructor(
