+++ app/src/main/java/ir/mehdiyari/krypt/ui/voice/record/RecordVoiceViewModel.kt
@@ -12,8 +12,8 @@ import ir.mehdiyari.krypt.files.logic.repositories.api.FilesRepository
 import ir.mehdiyari.krypt.files.logic.utils.FilesUtilities
 import ir.mehdiyari.krypt.ui.voice.recorder.SecondToTimerMapper
 import ir.mehdiyari.krypt.ui.voice.recorder.VoiceRecorder
-import ir.mehdiyari.krypt.ui.voice.recorder.meta.AudioMetaData
-import ir.mehdiyari.krypt.ui.voice.recorder.meta.AudioMetaDataJsonParser
+import ir.mehdiyari.krypt.voice.shared.entity.meta.AudioMetaData
+import ir.mehdiyari.krypt.voice.shared.entity.meta.AudioMetaDataJsonParser
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.Job
 import kotlinx.coroutines.cancelChildren
