+++ app/src/main/java/ir/mehdiyari/krypt/utils/Utils.kt
@@ -21,14 +21,3 @@ fun formatSize(size: Long): String {
 
     return "$internalSize $suffix"
 }
-
-fun Long.convertToReadableTime(): String {
-    val hours = this / 3600
-    val minutes = (this % 3600) / 60
-    val remainingSeconds = this % 60
-    return if (hours > 0) {
-        String.format("%02d:%02d:%02d", hours, minutes, remainingSeconds)
-    } else {
-        String.format("%02d:%02d", minutes, remainingSeconds)
-    }
-}
