+++ features/voice/player/build.gradle.kts
@@ -7,7 +7,11 @@ android {
 }
 
 dependencies {
-
+    implementation(project(":shared:files:files-logic"))
+    implementation(project(":shared:files:files-data"))
+    implementation(project(":shared:cryptography"))
+    implementation(project(":features:voice:shared"))
+    implementation(libs.moshiKotlin)
     testImplementation(libs.mockk)
     testImplementation(libs.coroutinesTest)
     testImplementation(libs.junit)
