+++ features/voice/player/src/main/java/ir/mehdiyari/krypt/voice/player/Utils.kt
@@ -1,0 +1,12 @@
+package ir.mehdiyari.krypt.voice.player
+
+internal fun Long.convertToReadableTime(): String {
+    val hours = this / 3600
+    val minutes = (this % 3600) / 60
+    val remainingSeconds = this % 60
+    return if (hours > 0) {
+        String.format("%02d:%02d:%02d", hours, minutes, remainingSeconds)
+    } else {
+        String.format("%02d:%02d", minutes, remainingSeconds)
+    }
+}
