+++ features/voice/player/src/main/java/ir/mehdiyari/krypt/voice/player/views/MusicPlayerViewModel.kt
@@ -1,4 +1,4 @@
-package ir.mehdiyari.krypt.ui.voice.player
+package ir.mehdiyari.krypt.voice.player.views
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
@@ -7,8 +7,10 @@ import ir.mehdiyari.krypt.cryptography.api.KryptCryptographyHelper
 import ir.mehdiyari.krypt.file.data.entity.FileEntity
 import ir.mehdiyari.krypt.files.logic.repositories.api.FilesRepository
 import ir.mehdiyari.krypt.files.logic.utils.FilesUtilities
-import ir.mehdiyari.krypt.ui.voice.audios.entity.AudioEntity
-import ir.mehdiyari.krypt.ui.voice.recorder.meta.AudioMetaDataJsonParser
+import ir.mehdiyari.krypt.voice.player.api.KryptMediaPlayer
+import ir.mehdiyari.krypt.voice.player.entity.MusicPlayerEntity
+import ir.mehdiyari.krypt.voice.shared.entity.AudioEntity
+import ir.mehdiyari.krypt.voice.shared.entity.meta.AudioMetaDataJsonParser
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
