+++ features/voice/player/src/main/res/drawable/ic_media_pause.xml
