+++ features/voice/player/src/main/res/drawable/ic_music_logo.xml
